requirements.txt
python-telegram-bot
pynput